# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QueryGenie API
  description: |
    Enterprise-grade API for QueryGenie - AI-powered database assistant with multi-role collaboration.
    
    ## Authentication
    All API endpoints require authentication via JWT tokens obtained through NextAuth.js.
    
    ## Rate Limiting
    - **Free Tier**: 100 requests per 15 minutes
    - **Pro Tier**: 1000 requests per 15 minutes  
    - **Enterprise**: Custom limits
    
    ## Error Handling
    All errors follow RFC 7807 Problem Details format.
    
  version: 1.0.0
  contact:
    name: QueryGenie API Support
    url: https://querygenie.com/support
    email: support@querygenie.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.querygenie.com
    description: Production server
  - url: https://staging-api.querygenie.com
    description: Staging server
security:
  - BearerAuth: []
  - SessionAuth: []

# =============================================================================
# Paths
# =============================================================================
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List user workspaces
      description: Returns all workspaces the user has access to
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    
    post:
      tags: [Workspaces]
      summary: Create workspace
      description: Creates a new workspace (user becomes owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Created workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token

  # Common Parameters
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8
...
    WorkspaceRole:
      type: string
      enum: [owner, admin, editor, viewer]
      description: |
        - **owner**: Full access including deletion
        - **admin**: Manage connections, members, settings
        - **editor**: Create/execute queries, edit docs  
        - **viewer**: Read-only access
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
        ownerId:
          type: string
          format: uuid
        isActive:
          type: boolean
        memberCount:
          type: integer
        connectionCount:
          type: integer
        role:
          $ref: '#/components/schemas/WorkspaceRole'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateWorkspaceRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
...
    ErrorResponse:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type
        title:
          type: string
          description: A short, human-readable summary of the problem type
        status:
          type: integer
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Request trace ID for debugging
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
                  value:
                    description: The invalid value

  # Common Responses
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            type: https://querygenie.com/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: Valid authentication is required to access this resource
            instance: /api/workspaces/123
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
          example:
            type: https://querygenie.com/errors/validation
            title: Validation Failed
            status: 400
            detail: One or more fields contain invalid values
            instance: /api/workspaces
            timestamp: "2024-01-20T10:30:00Z"
            traceId: abc123
            errors:
              - field: name
                code: required
                message: Name is required
              - field: slug
                code: pattern
                message: Slug must contain only lowercase letters, numbers, and hyphens
                value: "Invalid Slug!"
...